pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "filter"
version = "0.1.0"
description = "LLM-Powered Kanban board with Docker sandboxing"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Seth", email = "seth@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "jinja2",
    "python-dotenv",
]

[project.scripts]
filter = "filter.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"